+++ docs/components/content/Avatar/AvatarBasic.vue
@@ -5,17 +5,17 @@
 <template>
   <Space direction="vertical" :size="20">
     <Space :size="20">
-      <Avatar :size="64" />
-      <Avatar size="large" />
-      <Avatar />
-      <Avatar size="small" />
+      <Avatar :size="64" icon="i-carbon:user" />
+      <Avatar size="large" icon="i-carbon:user" />
+      <Avatar icon="i-carbon:user" />
+      <Avatar size="small" icon="i-carbon:user" />
     </Space>
 
     <Space :size="20">
-      <Avatar shape="square" :size="64" />
-      <Avatar shape="square" size="large" />
-      <Avatar shape="square" />
-      <Avatar shape="square" size="small" />
+      <Avatar shape="square" :size="64" icon="i-carbon:user" />
+      <Avatar shape="square" size="large" icon="i-carbon:user" />
+      <Avatar shape="square" icon="i-carbon:user" />
+      <Avatar shape="square" size="small" icon="i-carbon:user" />
     </Space>
   </space>
 </template>