+++ packages/components/Icon/src/index.ts
@@ -1,3 +1,2 @@
 export { default as IconArrowDown } from './ArrowDown.vue'
 export { default as IconEmpty } from './Empty.vue'
-export { default as IconUser } from './User.vue'