+++ packages/components/Avatar/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @lite-space/avatar
 
+## 0.0.1
+
+### Patch Changes
+
+- @lite-space/icon remove IconUser
+  @lite-space/avatar update
+
 ## 0.0.0
 
 ### Patch Changes