+++ packages/components/Icon/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @lite-space/icon
 
+## 0.1.1
+
+### Patch Changes
+
+- @lite-space/icon remove IconUser
+  @lite-space/avatar update
+
 ## 0.1.0
 
 ### Minor Changes